Fixes the underflow bug (see BUGS file for details).
git-svn-id: https://phycas.svn.sourceforge.net/svnroot/phycas/trunk@1031 b774380d-c90d-0410-997c-aec342e5d6fd

+++ phycas/src/cond_likelihood.hpp
@@ -26,7 +26,7 @@
 namespace phycas
 {
 typedef double LikeFltType;
-typedef unsigned UnderflowType;
+typedef long UnderflowType;
 
 /*----------------------------------------------------------------------------------------------------------------------
 |	Manages a conditional likelihood array for one end of an edge.
@@ -58,9 +58,9 @@ class CondLikelihood
 		LikeFltType *				cla;								/**< Pointer to conditional likelihood array stored by `claVec'  */
 		std::vector<LikeFltType>	claVec;								/**< Each element contains the likelihood conditional on a particular state, rate and pattern */
 
-		UnderflowType				uf_sum;								/**< Stores sum of underflow corrections over all sites */
-		UnderflowType *				uf;									/**< Pointer to the underflow correction array stored by `underflowExponVec'  */
-		std::vector<UnderflowType>	underflowExponVec;					/**< Stores log of the underflow correction factor for each pattern */
+		UnderflowType				uf_sum;								/**< Stores sum of underflow corrections over all sites. Used if SimpleUnderflowPolicy is in effect */
+		UnderflowType *				uf;									/**< Pointer to the underflow correction array stored in `underflowExponVec'. Used if PatternSpecificUnderflowPolicy is in effect */
+		std::vector<UnderflowType>	underflowExponVec;					/**< Stores log of the underflow correction factor for each pattern. Used if PatternSpecificUnderflowPolicy is in effect */
 
 		unsigned 					numEdgesSinceUnderflowProtection;	/**< The number of edges traversed since the underflow protection factor was last updated */
 	};
